fix: 순환참조 문제 해결
